+++ modules/AdminToolbox.MSGraph/Public/Get-msgMFAStatus.ps1
@@ -11,6 +11,10 @@
 
 function Get-msgMFAStatus {
 
+    [CmdletBinding()]
+    Param (
+    )
+
     #Confirm pre-requisites are met.
     Test-MSGraphRequirements -scopes 'Reports.Read.All'